Pose Library: do a broader search for the asset browser
When creating a pose, it gets assigned to the catalog that is currently
shown in the asset browser. To find which asset browser to use for this,
if the current window does not contain one, other windows are considered
as well. Before this change this was limited to the current screen.