Pose Library: refresh dopesheet & 3D view when changing asset library
The Create Pose Asset (`POSELIB_OT_create_pose_asset`) operator stores
the created asset in the current blend file. To avoid confusion, its
poll function refuses the use of the operator when the asset browser is
showing a different asset library. This means that the panels containing
a button for this operator need to be refreshed whenever another asset
library is selected.