Fix T92530: Rigify is trying to copy addon properties
The copy_custom_properties() function needs to check if a property is actually a custom property created by the user, or a property defined by an addon. I think we don't want to copy addon-defined properties here, since that's not what is usually meant by "custom property".

Reviewed By: angavrilov

Maniphest Tasks: T92530

Differential Revision: https://developer.blender.org/D13084