Copy Attributes: Copy Custom Properties
This patch adds the functionality to copy custom properties from active to selected bones, or from active to selected objects.

Test file: {F11494335}

It is currently not possible to copy custom properties from a bone to an object or vice versa, which is consistent with the fact that it is not possible to do this with constraints. Maybe in future.

Reviewed By: bassamk, sybren

Differential Revision: https://developer.blender.org/D11299