Pose library: remove unused parameter
Remove an unused `context` parameter from
`pose_library.pose_creation.create_pose_asset()`. Normally this wouldn't
make it into the release branch, but this particular function could be
the basis for custom pose library creation code, and thus it's good if
it has a nicer API in the release.

No functional changes.