Pose Library: explain why the 'for keymap' operators are there
The for-keymap version of the apply/blend operators take the Window
Manager's `poselib_flipped` property, and passes it to the regular
apply/blend operator. This makes it possible to bind a key to the
operator and still have it respect the global "Flip Pose" checkbox.

This behaviour was already in, this commit just adds a comment that
explains this.

No functional changes.