Fix T93509: Unintended navigation on Valve Index
The input bindings for "grab" navigation on the Valve Index controllers
could previously trigger the action even when the user did not intend
to do so (i.e. their hand was not actively gripping the sensor). This
was resolved by changing the bindings from "/input/squeeze/value" to
"/input/squeeze/force" and increasing the input threshold from 0.3 to
0.5.

Thanks to sybren for testing.