New animation add-on: Copy Global Transform
This add-on allows animators to copy the global transform of the active
object or pose bone onto the clipboard. It can then be pasted in three
different ways:

- To the current transform of the selected object/pose bone (could be a
  different one than was used for the copying).
- To selected keyframes.
- Baking to all frames between the first and last selected keyframe
  (defaulting to preview range or scene range).

All three methods are compatible with auto-keying.

The latter two methods *require* auto-keying to be enabled, to give the
animator control over which keying set to use, etc.

An earlier version of this add-on was used by the Blender Animation
Studio during Sprite Fright. Since then the two paste-to-frame-range
options were added, by request of the animators.

Reviewed by: campbellbarton

Differential Revision: https://developer.blender.org/D13570