Fix T94264: Pose Library, doubleclick-to-apply only in asset browser
Fix a keymap collision between the pose library and the file manager.
Only allow `applying a pose by doubleclicking it in the asset browser, and
not in the file browser. This prevents a crash; appending an Action
would temp-load it in order to apply it as a pose, which of course
conflicts with the "Append" functionality.