Fbx IO: improve export speed of 'rested' armatures
As part of Fbx Export, while handling Armature modifiers for Meshes, each
armature's position is backed up, then put into REST position for exporting,
and then restored back to original position. A dependency graph update is
triggered at the end of this.

This commit avoids the whole backing position setting + depsgraph update in
case the armature is already in rest position.

As an example, a model which I am developing for a game used to take
20 minutes for the Fbx Export. After this change, it only takes 20 seconds.

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D13712