Another way to disable old obj exporter.
The previous commits to split the obj importer/exporter into two
had the bad side effect of breaking the Python API for the importer
and exporter. So it was reverted and this less drastic method is used:
just don't put the python exporter into the menu. This way the
old scripts using bpy.ops.import_scene.import and .export will still
work (at least until we remove them completely in a future release).
The new obj exporter can be accessed in python right now as
bpy.ops.wm.export_obj(), which is probably not the best name but that
can be fixed later.