Fbx Import: Fix redundancy in action names
This commit aims at fixing the (very) redundant action names when importing animation data from Fbx files.

By default importing actions from an animated Fbx skeleton file leads to this:
**skeleton_name|skeleton_name|action_name|skeleton_name|action_name**

It works this way: **id_data.name|stack_name|layer_name**
These containers are generally made of the exact same strings multiple time.

This is a proposal to avoid such redundancy by stripping the "layer_name" in case it's identical to the "stack_name"
I'm not sure yet if one of them could be always skipped, I haven't experienced cases when they are different.

With this patch actions are named:
**skeleton_name|skeleton_name|action_name**

It could still be improved, since the "id_data.name" is generally the armature name, already contained in the "stack_name". But sometimes it's not (for Shape Keys for example: Key|skeleton_name|action_name)
Maybe more checks could be added to avoid redundancy even more. Just to open the discussion about it...

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D14130