Group FCurves by bone name when importing from FBX or BVH
When importing an animation from FBX or BVH the fcurves are currently shown as a very long list in the dope sheet.

When you manually create a keyframe they are grouped by bone name, which is much more user friendly. This patch groups imported animations by bone name too.

The changes are trivial:

* In the FBX case it was falsely using the object name to group all curves rather than the bone name.
* The BVH importer simply wasn't using the grouping feature at all.

Reviewed By: mont29

Maniphest Tasks: T82472

Differential Revision: https://developer.blender.org/D11269