VR: Add "Camera Landmark from Session" operator
Creates a new camera and "Custom Object"-type landmark from the VR
headset pose.

In contrast to the existing "Landmark from Session" operator that only
saves the headset rotation around the global z-axis, this preserves the
exact rotation of the headset by assigning it to the newly-created
camera (although the landmark itself still only inherits the z-rotation
component).