Fix .X3D/.WRL importer crashing with empty IndexedFaceSets
When importing a .x3d file containing an empty IndexedFaceSet, like this for example:

```
Shape {
  geometry IndexedFaceSet {
    coord Coordinate {
      point [

      ]
    }
    coordIndex [

    ]
  }
}
```
`import_x3d.py` throws an exception, because `x_min`, `x_max`, ... are not initialized/still set to `None`.
This fixes the issue by initializing the mentioned variables to `inf`/`-inf` respectively and also further simplifies the code by utlizing the `min` and `max` builtins.

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D14470