Themes: Adapt themes to the dot grid background
Update the themes to the recent changes made to the node editor's dot
grid background:

1. Make sure themes aren't using a "Grid Levels" value of greater than 3

2. A lot of themes weren't updated to the dot grid at all, yet, and
still used only 2 "Grid Levels" leading to the old dot grid being hardly
visible. Those are changed to 3 "Grid Levels", as well, which most
closely matches the visual density of the old line grid.

3. Additionally the "Grid" colour of the "Deep Grey" theme is darkened
a bit to account for the dot grid being colored differently than the old
line grid, which lead to the dot grid being invisible.