Fix T98840: Recursion Error with Development - Edit Operator Addon
In the prior versions of this addon I had to generate a list of builtin and standard python modules,
which need to be excluded in the walk function.

This function is being called recursively over all blender py modules
and addon modules.

Certain builtin or standard modules of python seem to have some recursion in their module structure,
which needs to be avoided.

Since Python 3.10, you can use 'sys.stdlib_module_names' for a basic list.
Possibly the old method generated an incomplete list.