AnimAll: add attribute animation
Vertex colors have been replaced by a more generic attribute system,
so we need to support that.

This allows animating not only color attributes, but also other types.
Not sure that's very useful as most attributes can't be manually
edited as far as I understand, but it *is* now possible!