AnimAll: remove vertex colors
Vertex color access through `Mesh.vertex_colors` is deprecated and
will be removed eventually.
https://wiki.blender.org/wiki/Reference/Release_Notes/3.2/Python_API#Deprecation

It is already not entirely functional for AnimAll's workflow: if there
are multiple vertex color layers, the active one cannot be selected
because the interface is no longer exposed. This results in the user
being unable to select which layer they want to animate.

Besides, all functionality is covered by the more generic attributes
system. An operator is provided to upgrade animation curves still
using the old system.