Depsgraph Debug: add an operator to generate an SVG and open in browser.
Rather than converting the dependency graph to a bitmap image,
it is more advantageous to use the vector SVG format and open it
in a web browser (any modern browser supports the format), because
it allows text searches within the image.

This patch adds a new operator and button that does just that.

Differential Revision: https://developer.blender.org/D15325