Rigify: fix T95864 - inconsistent bone roll in the wolf metarig.
Edit mode represents bone orientation via their head and tail
locations, plus a roll value. This math has a singularity in
the -Y direction, resulting in instability and discontinuity
for bones exactly aligned to the Y axis (the tail flips the
deform bones, so +Y is also bad).

The only reliable way to avoid problems is to un-align the
bones via very small offsets in the Z direction.