Amaranth: Add check for unsaved images in Save/Reload operator
The regular Save operator does not automatically save changes in images,
users become aware of this when quitting Blender or changing files.

This became a problem for the Save/Reload operator since you'd lose work
without noticing it. This commit adds a check and let the user now in the
terminal which image files have not been saved.

Fixes T73905