Fix T96194: Node Wrangler: use enabled sockets for lazy mix
Add utility function to check for the first output that is actually enabled
rather than always using the first output of a node.

The previous behavior would lead to confusing situations, where a mix/math
node was added but didn't seem to be connected, when the first output was
currently disabled.

Differential Revision: https://developer.blender.org/D14251