Fix T101708: Bug when joining nodes with thick lines
Caused by D15961, which changed the behavior of `dpi` and
`pixel_size`.

Instead of using a single value for `dpifac` (renamed to `dpi_fac`),
use a value depending on whether we want pixel density or line
width.

This patch has the side-effect that the line width from the user
preferences is now respected. It seems more logical that way, but I
can't test with a hi-dpi display to see how that looks.

If we want to get the old behavior back, we can just get rid of
`prefs_line_width()` and replace its calls with `dpi_fac()`.

A similar issue was also fixed for Icon Viewer.