Rigify: port the ability to generate Action constraints from CloudRig.
CloudRig has a feature that allows the user to automatically generate
Action constraints that move bones of the rig based on the position
of other bones. This is done by adding and configuring the actions
in a UI panel of the metarig. The feature also supports corrective
actions that activate based on the state of two other actions.

This ports the feature to base Rigify with the necessary changes in
code organization and style, and replacing CloudRig-specific code.

There are also some functional changes:

- The order of action constraints is reversed.
- The way symmetry of LOCATION_X is handed is changed to
  match how Paste Pose Flipped works.
- The action slot UI is shown even without a generated rig.
- More alerts in the UI, e.g. for duplicate rows.

Differential Revision: https://developer.blender.org/D16336