Fix T102396: ValueError: matrix does not have an inverse
Refactor the matrix stack in a way that does not require matrix
inversion. Basically, store the state of the final transform in
the stack.

Technically this makes regression test to fail with Blender Icons,
but the new code gives more correct icons. So the reference image
is to simply be regenerated.