Official add-ons: fix printf-style format translation for i18n
Following up to D15615, I noticed that many messages are either
properly extracted, but not translated due to the use of printf-style
formatting in strings, or not extracted at all. Using the pgettext
function explicitly before formatting fixes that.

Since the affected messages are already extracted, no additional work
is needed from translators.

Reviewed By: mont29

Differential Revision: https://developer.blender.org/D16373