Fix T99623: exception in pose library when the path to the library is missing
Add check if `context.copy` is not `None`, before trying to call it.

When the path to the asset library points to a non-existing directory,
`context.copy` is set to `None`; when the code reaches `return
bpy.ops.poselib.apply_pose_asset.poll(context.copy())` in the `poll`
method of the class `POSELIB_OT_apply_pose_asset_for_keymap`, it throws
an exception since `None` is not callable.

Reviewed By: sybren

Maniphest Tasks: T99623

Differential Revision: https://developer.blender.org/D15427