Pose Library: Usability pass/redesign
This implements a redesign on how pose library is exposed to the
artists, unifying the experience within other assets.

We now try to constraint the different usages to specific editors,
and optimize the experience in each editor to those minimum requirements.

In simple terms this means to encourage the creation of assets in the
Action Editor or the Asset Browser, using assets in the Viewport.
Having the main functionality available at the click of a button,
and the advanced options available via (context) menus.

For the complete design see: T103265

------------------------------------------------

This implements:
* Viewport:
 - Remove Create Pose Asset.
 - Remove Copy Pose As Asset.
 - Remove Flip Pose checkbox.

* Asset Browser:
 - Add "Paste as New Asset" and "Create Pose Asset" to a new Asset menu.
 - Remove Pose Library panels.

* Context Menu:
 - Add Apply Pose Flipped.
 - Add separators.

* General:
 - Rename Paste/Copy As to Paste/Copy as

Follow ups:
 * Asset Browser: New catalog menu (D16820)
 * Asset Shelf (T102879)
 * Show shortcuts in the Pose Assets context menu (T103267)
 * "Update" operators (Paste as Update and Update Pose Asset)
 * Action Editor: Catalog pop-up for new created poses
 * Action Editor: Bring the Pose Library panel to the Animation tab

The update operators are welcome, but outside of the scope of the
usability change. If/when they are implemented, they will fit within the
Edit menu.

The Action Editor changes should be tackled shortly.

Differential Revision: https://developer.blender.org/D16795