Sun Position: add Show Surface and Show Analemmas options
Sometimes during the design it's difficult to rapidly visualize the
Sun trajectory for a given location in the viewport.

This patch adds the possibility to visualize the sun path surface and
analemmas for each hour in the viewport.

Currently there is an option to create a collection with objects to
draw analemmas and diurnal ready for render but these new options
options give fast viewport feedback, something like the Show North
option.

Reviewed By: pioverfour

Differential Revision: https://developer.blender.org/D16936