Fix T103976: FBX Export: Duplicate materials cause material indices to be offset
FBX exporter would add duplicate materials to a mesh only once, but would increment the material index once for each duplicate, causing material indices to become offset and/or go out of bounds.

This patch adds a check that skips incrementing the material index if the material has already been added to the mesh.

---

I've added an 'XXX' comment that points out that this area of code causes material slots on the same mesh that contain the same material to be merged into one material slot because the `material_indices` dict's keys are materials only, so a material can only be given one index (this behaviour is not changed by this patch).

I find this to be undesirable behaviour because material slots are useful for working with different parts of the same mesh, such as animating the material of a specific material slot in external software.

Modifying the code to support maintaining multiple slots with the same material is beyond the scope of this patch and I haven't looked into what modifications would be required beyond changing the keying or changing the indices to be multivalued. There could even be a restriction that the FBX format doesn't allow duplicate materials on a mesh in the first place, which would likely require creating copies of duplicate materials on export instead.

Reviewed By: mont29

Maniphest Tasks: T103976

Differential Revision: https://developer.blender.org/D17052