Fix T75982: FBX export with empty material swapps materials around
FBX exporter code was creating the list used for mapping material indices to FBX indices from only the materials being exported.

Since there is no material for empty material slots, the list used for mapping would be reduced in length by one for each empty material slot.

The mesh's material indices would remain unaffected by the reduction in length of the mapping list, causing any material indices after an empty material slot to be mapped to the wrong FBX index.

This patch changes the creation of the mapping list so that any material slots with a material that is not being exported are mapped to the default material instead of being excluded.

Reviewed By: mont29

Maniphest Tasks: T75982

Differential Revision: https://developer.blender.org/D17034