Fix T103949: FBX Importer/Exporter: Edge creases missing when re-importing some exported meshes
FBX exporter was missing the edge of the last loop of meshes where that last edge wasn't used by any other polygons due to no end for the last polygon being specified.

This patch adds the missing end for the last polygon so that the export of the last edge is never missed.

---

The reason the edge creases would be missing was because the exported .fbx were malformed, containing one less edge (and thus one less edge crease) than they should.

The missing edges would not be physically missing from the mesh upon import, presumably due to edges in .fbx being represented as loops, of which, none were missing.

This patch does not address the observation that in Blender 3.4 the import of a .fbx malformed in this way results in no creases being imported as opposed to in 3.3 where all the creases that are actually in the .fbx do import (though of course one is missing from the exported .fbx itself).

Reviewed By: mont29

Maniphest Tasks: T103949

Differential Revision: https://developer.blender.org/D17040