FBX Export: Speed up shape key weights and normals.
* Add `break` statement once `vg_idx` is found.
* Get `me.vertices` outside the main loop.
* Replace lists with arrays to skip the step of converting from list to array
  in `elem_data_single_float64_array`.

No changes are made to the exported file with this patch.

The `break` statement is the main speedup in most cases.
The other changes only result in about a 5-10% speedup (observed by comparing
cases where vg_idx is not in the vertex's weights, so all weights on the
vertex have to be iterated).

Differential Revision: https://developer.blender.org/D17168