Node Wrangler: apply autopep8 formatting
The node wrangler source code wasn't consistently formatted before,
which made it quite annoying to make changes. This formats it with the
same formatter used for other python files in blender's source code, so
that, in principle, the IDE would already be configured correctly for
anyone working on blender.

Pull Request #104463