Assign modifiers to variables instead of getting them by name
Using hardcoded names to retrieve modifiers can result in errors if
the user has enabled UI translation, because the new modifiers may not
have the expected English names if the UI is translated.

So, use the data API to create the modifiers instead of the ops API,
and assign nodes to variables instead of getting them by name.