Fix #104438 : Lazy mix connects wrong sockets
Because the new mix nodes have different modes, the correct output
socket to connect is not just the first one, but the first enabled one.
There were already some checks in place for this, but apparently not yet
in all the needed places.

The current state of the code for the "Merge Nodes" operator is
terrible, and should probably be refactored in a future update. It is
becoming very hard to maintain.

Pull Request #104457