Speed up FBX export of polygon indices and edges with numpy
Use buffers matching the C types of the data in foreach_get to avoid having to iterate and cast every single element in the C foreach_getset function.

Get the edge index for each exported edge key and use this to map per-edge data, such as creases, to the exported edges.

~9-17 times faster for a minimal export (no loose edges, smoothing or creases exported)
~16-23 times faster for a maximal export with FACE smoothing (loose edges, face smoothing and creases exported)
~32-40 times faster for a maximal export with EDGE smoothing (loose edges, edge smoothing and creases exported)

Isolated parts:
~5-7 times faster loose edges processing
~2-4 times faster face smoothing export
~73-96 times faster edge smoothing export
~62-75 times faster creases export

In the rare case where there exist multiple edges with the same edge-key, this patch can result in a different export because the single edge exported from those multiple edges will have the per-edge data of the first found edge in loops. Before this patch, the per-edge data would be from whichever edge with the same edge-key was iterated last since it would overwrite the previously set per-edge data of the exported edge.

-----
Polygon indices and per-edge data are interconnected so had to be part of one patch. This covers the export of polygons and edges, loose edges, edge/face smoothing and edge creases for subdivision.

The `t_lvi` array used in this patch could be re-used in the export of UVs and/or Vertex Colors since the patches for both of those areas require the same data albeit without any extra elements that have been added for loose edges. As a note on doing this `t_pvi` array may currently be a view of `t_lvi`, but for `t_lvi` to be re-used for UVs/Vertex Colors, `t_pvi` would have to always be a created as a copy of `t_lvi` so that modifying `t_pvi` does not also modify `t_lvi`. This can be achieved by replacing `t_pvi = astype_view_signedness(t_lvi, pvi_fbx_dtype)` with `t_pvi = t_lvi.astype(pvi_fbx_dtype)`
See https://projects.blender.org/blender/blender-addons/pulls/104453 and https://projects.blender.org/blender/blender-addons/pulls/104454

Co-authored-by: Thomas Barlow <github@mysterymayhem.co.uk>
Pull Request #104451