Speed up FBX export of shape keys with numpy
Use buffer matching the C type of the "co" property in foreach_get to avoid having to iterate and cast every single element in the C foreach_getset function.

Replace use of vcos_transformed_gen mesh transform helper with numpy version.

Only get cos of shape keys that are needed, since shape keys that are relative to themselves and have no other shape keys are relative to them can be skipped.

~7-12 times faster for shape keys that are entirely equal
~11-25 times faster for shape keys that are entirely different and not similar
~16-28 times faster for shape keys that are almost entirely different but similar

This patch does usually slightly change the exported file because the math for calculating the difference between a shape key and its relative key is done with float64 precision, matching the fbx export type. Beforehand, the difference would be calculated using mathutils.Vector which are single precision float (usually float32).

Because numpy.isclose is not symmetrical, this patch can also rarely result in a difference in which shape key cos are considered similar and are therefore not included in the export.
Consider the relative difference between 0.5 and 1.0:
If 1.0 is considered the reference value there is a 50% difference, but if 0.5 is considered the reference value there is a 100% difference.
math.isclose and fbx_utils.similar_values_iter always picks the value with greater magnitude as the reference value, whereas numpy.isclose picks the second argument as the reference value.
math.isclose(0.5, 1.0, rel_tol=0.5) => True
math.isclose(1.0, 0.5, rel_tol=0.5) => True
numpy.isclose(0.5, 1.0, rtol=0.5) => True
numpy.isclose(1.0, 0.5, rtol=0.5) => False
The relative key of each shape key is used as the reference value for numpy.isclose.

-----
Whether geom_mat_co is set or None didn't make much difference to the total speedup, which is why the timing comparison in the commit message doesn't mention geom_mat_co.

The speedup around shape key weights is very minor. Without the use of accessing the numpy arrays through their underlying memoryview objects, the main loop would be slower since numpy arrays seem to be quite a bit slower to iterate and index one element at a time compared to Python arrays.

It might be possible to speed up the export of shape key weights by figuring out ahead of time all the indices of vertices that are needed for shape key weights so that each vertex is only ever indexed at most once, but this sounded very complicated and I don't think shape key weights typically get used much. I'd rather hope and wait for some way to quickly access vertex group weights, similar to foreach_get/set, to get added to the Python API.

Co-authored-by: Thomas Barlow <github@mysterymayhem.co.uk>
Pull Request #104452