FBX export: Option to prioritize active color at export.
Some other software discard other color attributes besides the first one -
so the option to prioritize active_color to be exportred first will be
helpful in these cases. It is especially important since currently there
is no option to move color attributes around in Blender.

Explained it in details here:
https://projects.blender.org/blender/blender/issues/104827

To solve that I've added that option in FBX exporter.